修改丢包统计和码率统计的计算方式

+++ estimator/rate_stat.h
@@ -31,7 +31,7 @@ typedef struct
 
 	float			scale;
 
-	int				accumulated_count;
+	int64_t			accumulated_count;
 	int				sample_num;
 }rate_stat_t;
